fix(Calendars): Fix calendars to update in real time
Switched order of Query and Calendar component to pass down data
